requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
gunicorn==21.2.0 